app/requirements.txt
uvicorn
fastapi
pydantic

singleton_package
json-advanced
python-dotenv
debugpy
ipython

beanie
redis
